Switch ownership to ecommerce-experience
As discussed initially with Marcus and then clarified with Jen Page and
Draper, this team seems to make the most sense as an owner for
lunar-core.

Even with plans to deprecate it on the horizon, this should be owned by
a team with frontend experience and it's used exclusively in github
projects owned by this domain.